+++ skytemple_files/list/actor/model.py
@@ -52,7 +52,7 @@ class ActorListBin(Sir0Serializable):
     def serialize(self) -> bytes:
         return self.sir0_serialize_parts()[0]
 
-    def sir0_serialize_parts(self) -> Tuple[bytes, List[int], Optional[int]]:
+    def sir0_serialize_parts(self) -> Tuple[bytes, List[u32], Optional[u32]]:
         string_codec.init()
 
         out_data = bytearray()
@@ -72,7 +72,7 @@ class ActorListBin(Sir0Serializable):
             entry_buffer = bytearray(LEN_ACTOR_ENTRY)
             write_u16(entry_buffer, entry.type, 0)
             write_u16(entry_buffer, entry.entid, 2)
-            sir0_pointer_offsets.append(len(out_data) + 4)
+            sir0_pointer_offsets.append(u32(len(out_data) + 4))
             write_u32(entry_buffer, pointer_offsets[i], 4)
             write_u16(entry_buffer, entry.unk3, 8)
             write_u16(entry_buffer, entry.unk4, 10)
@@ -82,8 +82,8 @@ class ActorListBin(Sir0Serializable):
         self._pad(out_data)
 
         # 4. Write sub-header
-        data_pointer = len(out_data)
-        sir0_pointer_offsets.append(len(out_data))
+        data_pointer = u32(len(out_data))
+        sir0_pointer_offsets.append(u32(len(out_data)))
         out_data += pointer_data_block.to_bytes(4, byteorder="little", signed=False)
         out_data += len(self.list).to_bytes(4, byteorder="little", signed=False)
 
@@ -93,7 +93,7 @@ class ActorListBin(Sir0Serializable):
     def sir0_unwrap(
         cls,
         content_data: bytes,
-        data_pointer: int,
+        data_pointer: u32,
         static_data: Optional[Pmd2Data] = None,
     ) -> "ActorListBin":
         return cls(content_data, data_pointer)
